chore: rename Relay
