+++ src/core/internal/modes/relay.ts
@@ -12,7 +12,7 @@ import { waitForCallsStatus } from 'viem/actions'
 import * as Account from '../../Account.js'
 import * as Key from '../../Key.js'
 import type * as Porto from '../../Porto.js'
-import * as Relay from '../../Relay.js'
+import * as Relay from '../../Rpc.js'
 import type * as Storage from '../../Storage.js'
 import * as Call from '../call.js'
 import * as Delegation from '../delegation.js'
