+++ src/core/internal/viem/actions.test.ts
@@ -7,7 +7,7 @@ import * as Anvil from '../../../../test/src/anvil.js'
 import { exp1Abi, exp1Address } from '../../../../test/src/porto.js'
 import { getPorto } from '../../../../test/src/porto.js'
 import * as Key from '../../Key.js'
-import { sendCalls } from '../../Relay.js'
+import { sendCalls } from '../../Rpc.js'
 import type * as Capabilities from '../rpcServer/typebox/capabilities.js'
 import {
   createAccount,
