+++ src/index.ts
@@ -5,6 +5,6 @@ export * as Key from './core/Key.js'
 export * as Messenger from './core/Messenger.js'
 export * as Mode from './core/Mode.js'
 export * as Porto from './core/Porto.js'
-export * as Rpc from './core/Relay.js'
+export * as Rpc from './core/Rpc.js'
 export * as RpcSchema from './core/RpcSchema.js'
 export * as Storage from './core/Storage.js'
